Restart wartende/fehlgeschlagene Importe im Import-Dialog
Bislang blieb ein Import, dessen Tab geschlossen oder neu geladen wurde,
während er noch "Wartet…" war, für immer in diesem Status hängen: die
Verarbeitungs-Queue lebt nur im Arbeitsspeicher des Tabs und wurde nie
aus Firestore wiederhergestellt.

Jetzt wird pro Job eine restartfähige Quelle (URL bzw. komprimierte
Bilder/Text) in Firestore mitgespeichert, sodass "Neu starten" den Import
jederzeit erneut anstoßen kann – auch nach einem Reload. Die Run-Funktionen
der drei Import-Wege (Web, Universal, Foto-Scan) wurden dafür aus den
Modals in ein gemeinsames src/utils/importRunners.js verschoben.
